pyproject.toml
[tool.poetry]
name = "selenium-killer"
version = "0.1.0"
description = ""
authors = ["Rodrigoneal <rodrigho2006@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
httpx = "^0.27.0"
typing-extensions = "^4.9.0"
bs4 = "^0.0.2"
capmonstercloudclient = "^1.3.4"
python-dotenv = "^1.0.1"
requests = "^2.31.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.1"


[tool.poetry.group.linter.dependencies]
ruff = "^0.2.2"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
